+++ packages/client-library-otel/src/utils/index.ts
@@ -1,5 +1,9 @@
 export { getFromEnv, type FpHonoEnv } from "./env";
-export * from "./errors";
+export {
+  errorToJson,
+  isLikelyNeonDbError,
+  neonDbErrorToJson,
+} from "./errors";
 export { isUintArray } from "./is-uint";
 export { safelySerializeJSON } from "./json";
 export {
@@ -9,22 +13,5 @@ export {
   cloneRequestForAttributes,
 } from "./attributes";
 export { isWrapped } from "./wrapper";
-
-export function isPromiseLike<T>(value: unknown): value is PromiseLike<T> {
-  return (
-    isObject(value) &&
-    "then" in value &&
-    typeof (value as { then: unknown }).then === "function"
-  );
-}
-
-export function isObject(value: unknown): value is object {
-  return typeof value === "object" && value !== null;
-}
-
-export function objectWithKey<T extends string>(
-  value: unknown,
-  key: T,
-): value is { [K in T]: unknown } {
-  return isObject(value) && key in value;
-}
+export { isObject, objectWithKey } from "./object";
+export { isPromiseLike } from "./promise";
