+++ packages/client-library-otel/src/utils/object.ts
@@ -1,0 +1,10 @@
+export function isObject(value: unknown): value is object {
+  return typeof value === "object" && value !== null;
+}
+
+export function objectWithKey<T extends string>(
+  value: unknown,
+  key: T,
+): value is { [K in T]: unknown } {
+  return isObject(value) && key in value;
+}
